backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
notion-client==2.4.0
openai==1.79.0
chromadb==1.0.20
sentence-transformers==5.1.0
lingua-language-detector==2.1.1
torch==2.8.0
numpy==2.3.2
emoji==2.14.1
tiktoken==0.12.0
gunicorn==21.2.0